RSP-653 [Accessibility] FormItem without label prop should not generate empty label tag (#455)
For some use cases, it may be desirable to have a FormItem with content aligned with other fields in the Form table layout but with no label text.

One such example might be a Checkbox to accept licensing terms and conditions. The checkbox has its own label.

In this case, the `.spectrum-Form-itemLabel` would be empty and as such, it should not be rendered using a `label` tag with a `for` attribute. The empty `.spectrum-Form-itemLabel` should be rendered using a `div` tag.

Add `aria-labelledby` to `LabelBase` to support composite controls like ComboBox or RadioGroup.