[RSP-736] Add renderItem prop for custom menu item rendering for Autocomplete, ComboBox, SelectList, and Select (#488)
* RSP-736 [Accessibility] Autocomplete/Combobox: Throws error when label is HTML or React

A use case for a ComboBox from which to select a time duration, like “08:00” for “8 hours”. The challenge with this, from an accessibility standpoint, is to ensure that a screen reader announces the value as a duration, rather than as a time value like “8 o’clock”. The theoretically appropriate HTML element for this purpose is the <time> element using a valid ISO 8601 Duration as the datetime attribute. (I say theoretically, because assistive technology support for the <time> element does little more than announce the <time> element’s text content at this point, so adding an aria-label to the element is also necessary).

Fix by adding `renderItem` prop to specify a wrapper method for rendering the item label.