Eslint our jsdocs (#662)
* Docs eslint our jsdocs
We have some requirements that we've been submitting one off PR's for, ending a comment in a period, starting it with a capital letter. We might as well use the eslint plugin for jsdocs. I've added a few rules to get us started, we require descriptions, we require that those descriptions are complete sentences, and  the block to define them should be tabbed correctly.
I also check that the tags are valid jsdocs.

* remove some overly aggressive docs'ing

* small fixes to documentation

* missed a period

* fix lint

* Fix typo