Add specificity to flex gap children to fix docs (#886)
In general this is a problem with specificity, the component may be rendering a button and need to do a style reset. In that case, we are at the mercy of the order of css loading. We can raise that by one specificity level and that will solve most cases. Someone could override it pretty easily and accidentally. We might want to consider !important for this very specific use case.

Co-authored-by: Devon Govett <devongovett@gmail.com>